125/100

+++ libft/ft_strnew.c
@@ -17,7 +17,7 @@ char	*ft_strnew(size_t size)
 {
 	char *str;
 
-	str = (char *)malloc(((int)size + 1) * sizeof(char));
+	str = (char *)malloc((size + 1) * sizeof(char));
 	if (str == NULL)
 		return (NULL);
 	ft_memset(str, '\0', size + 1);
